% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{factorial}
\alias{factorial}
\title{Factorial}
\usage{
factorial(n)
}
\arguments{
\item{n}{is the number which you want the factorial of}
}
\value{
the factorial of the number n
}
\description{
Factorial
}
\examples{
factorial(5) = 120
factorial(3) = 6

}
